+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Fix Slider/Slideshow component autoplay in Firefox
 - Fix using input of type range in Modal component on touch devices
 - Fix accessibility in Slideshow/Slider and Nav component
+- Fix prevent initial transition in Sticky component
 
 ## 3.16.6 (March 10, 2023)